requirements.txt
asgiref==3.7.2
axios
certifi==2023.7.22
distlib==0.3.7
Django==4.2.5
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djoser
filelock==3.12.4
firebase
pipenv==2023.10.3
platformdirs==3.11.0
python-dotenv
setuptools==68.2.2
social-auth-app-django
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.5
wheel==0.41.2
